MM: basic paging and use `linked_list_allocator::LockedHeap`
TODO 1: Potential bug in paging where it's impossible to read the entire
4096 bytes long page.

TODO 2: Create own heap allocator

+++ kernel/units.rs
@@ -16,13 +16,13 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-pub const BYTE: u64 = 1;
+pub const BYTE: usize = 1;
 
 #[allow(non_upper_case_globals)]
-pub const KiB: u64 = BYTE * 1024;
+pub const KiB: usize = BYTE * 1024;
 
 #[allow(non_upper_case_globals)]
-pub const MiB: u64 = KiB * 1024;
+pub const MiB: usize = KiB * 1024;
 
 #[allow(non_upper_case_globals)]
-pub const GiB: u64 = MiB * 1024;
+pub const GiB: usize = MiB * 1024;
